[LTB-49] Remove manual connection handling
Manual handshake is too complex and it was removed in favour
of embedded ZMQ connection routines run in the background.
Previously we had ROUTER-ROUTER communication, and thus on client
we needed to store the map from server hosts (as we see them)
to servers' identities (random bytes server set). This effectively
lifted all the complicated details of manually handling the
connection to the high level.

Instead, this commit introduces a pair of sockets per peer (DEALER
and SUB). Nothing in the library interface changes, but inside
we now poll on N dealers plus N subs at the same time, and
when we get the message we know exactly where it came from.
This removes all the burden of manual connection handling
without any compromise.

In particular, it simplifies the following cases:
* How to connect to the client?
* How to deal with connection timeouts?
* What to do when the server changes internal id (because
  it has restarted?)

The (3) would lead to the bug in the previous code version actually,
so now we do not have this problem.