[LTB-49] Change data polling mechanism
Previously, data was polled using STM (in both client and server
brokers). This approach was slightly more complicated and error-prone.
Main broker should have polled on both sockets and some tvars,
so a concept of Adapter was introduced to wrap sockets into
STM-compatible variables, which in turn allowed us to use one
big STM poll. This approach showed itself to be complicated
and it also introduced weird bugs (DSCP-177) which led to ugly
workarounds.

This commit changes main polls to zmq based polls and does exactly
the opposite -- now we wrap STM actions so they are compatible with
the ZMQ poll. In order to do this, we introduce a concept of
InternalQueue, which is a pair of PAIR sockets plus a normal
TQueue. Whenever we (ZMQ) poll, we include the "out" pair socket
into the poll call, and whenever we send data through the internal
queue, we ping on this socket and transfer the data through the
tqueue, so it can be read after the ZMQ poll returns "In" event.
Listeners and clients are handled this way: we fork a small worker
which constantly reads from STM BiTQueues of listeners/clients
and pushes everything to the internal queue (which is then processed
by the main broker).