Implement config parsing from env
Problem: we already support parsing from yaml config and from CLI
options, but in the place where I'm currently working it is common to
use mostly environmental variables. Meanwhile, I failed to find any
decent solution for this which would do exactly environment parsing and
nothing more.
If we implement env parsing here, this can become the superior library
in the Haskell ecosystem solving this task.

Solution: implement env parsing.

Variable name is evaluted from the full path to the variable in config,
and values are parsed using the new dedicated typeclass.