Add conversion between datatype and config
Problem: it is often a case when a piece of configuration is a datatype
from a third-party library, would be nice to have a way to work with
them easily.

Solution: the most proper way seems to be having a type family which
assosiates a config to the given datatype, and methods to implement a
conversion from list of config Items to the datatype. This way we still
allow config values to be partially absent, and convert to the datatype
only after finalization.

First I tried to implement this via generics-sop, but utilities there
occured to be not very helpful in our case, besides, there field type
and field name are put into different places which is inconvenient.
In constrast, plain GHC.Generics provide field names and types next to
each other.