Add required env vars collector
Problem: unlike with parsing config or CLI parameters, with env
variables it is not always obvious which exact variables the config
expects. In case if name unexpected for a user is generated, this can be
difficult to debug.

Solution: provide a method which enlists all env keys expected by
configuration.