[LTB-61] Add ToJSON instance for config file
Problem:
We have FromJSON instance so we can parse configuration from valid
json or yaml. But there are cases when we want print this config file,
send, log, etc. So it's useful to have ToJSON instance too.

Solution:
Add ToJSON instance for ConfigRec, converting to json happens with
help of OptionsToJson class which recursively goes through config file
and convert values.

Note: we independently implemented `ToJSON` in this PR before we noticed #53