[LTB-65] Add Buildable for ConfigRec
Problem:
This was made here #53. But this PR became still.
It'll be good to proceed this PR and add Buildable instances, cause
it's quite useful.
We not going to merge #53, because Json instances was added recently
independent from this PR (we didn't know about it).

Solution:
Move Buildable instances (along with tests) form #53.